Bring back more limited LogT
This is different than the previous incarnation because:

* It doesn't take monadic versions of the context/namespace getters and
  thus sheds a lot of tricky instances.

* Does not do any sort of combining logic.

* Does not have to be ordered with BlankLogT as some sort of weird
  type-level list of contexts.

The main point of LogT now is for when you need to pull out your katip
context at a point in time and rebuild a shorter transformer stack to
run right over IO. A good example is if you're writing a library that
has to do something in IO (such as another thread, in an Async, etc.)
You want to pull out your log context and use it in IO without having to
roll your own transformer with Katip/KatipContext. For *applications* it
is still preferable to use a Reader-based approach where adding
namespaces and contexts is as simple as a call to `local`.