WIP: katip internal queueing
Doing this stopping point because I'm pretty sure things will get a
lot less messy if I do indeed just put the finalizer on the scribe
data structure itself. We can't get away with a non-breaking change
and it will make it much harder to forget or mishandle the finalizer.