Define exception behavior in finalizers
Before we were swallowing exceptions in the finalizer
wrapper. Instead, we let closeScribe throw, and we let closeScribes
run all finalizer before rethrowing the first error. The idea is that
if the user is closing a single scribe, they should be able to react
if there's an issue. If they're closing all scribes, we should make
sure resources are freed but then let them respond. I suppose the
other way would be to catch and return a list of exceptions associated
with the finalizer. I could be convinced of that API too.