Derive or implement `Show` and `Eq` instances
This commit adds `deriving` directives for `Show` and `Eq` to `LogStr`,
`ScribeSettings` and `ColorStrategy` where applicable and adds a manual
`Eq` instance for `Item a` (due to a missing `Eq` instance for `Loc` in
old versions of `template-haskell`).

Also add a basic test for the `Item a` `Eq` instance.

Fixes: Soostone/katip#69
See: https://github.com/Soostone/katip/issues/69