Update deprecated mapping specification
This was supposed to be deprecated in ES 5.x but I guess it got fully
removed in 6 and is no harm in 5. ES split the "string" mapping type
into unanalyzed "keyword" and analyzed "text". This updates the
mapping for v5 to explicitly differentiate and is more deliberate
about which fixed mapping fields are analyzed and which aren't.