Fix analyzed/unanalyzed bug in ESV1
When we started differentiating katip internal fields being analyzed
or not, I did not go back and propagate this to ESV1. This is a
typewise minor release but because it changes index semantics moving
forward, I will probably release as a major version. In effect, if you
use ESV1, this will update your mapping on setup and any new
indices (e.g. if you use daily indexing, tomorrow's logs) will
correctly stop analyzing various fields that are meant to be
tokens. Without doing this, it is difficult to filter logs precisely
on things like host if there's dashes or something in your hostnames.