Move ThreadId prefix out of JSON newtype and into handle scribe
My reasoning here is that it makes sense in the bracketed handle
scribe to differentiate these random looking numbers next to eachother
as PID and ThreadId, respectively. Moving it out of the newtype means
backends like elasticsearch and datadog don't log it wiith the extra
"ThreadId " prefix, which is unnecessary because the log data id
structured and identifies the field with the key name.