Try removing MonadIO superclass
This superclass isn't really needed by the Katip class
itself. Basically it prevents you from adding a MonadIO constraint if
you do use logging, but:

1. You get confusing redundant constraint errors when you also have a
MonadIO constraint.
2. It forces you to add MonadIO in some cases where you're not
actually logging.

I had 10 minutes to kill so I thought I'd see how hard it'd be to
remove.

This is in reference to #88